tests

+++ src/app/testimonials/testimonials.component.ts
@@ -1,5 +1,5 @@
 import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
-import { MatBottomSheet } from '@angular/material';
+import { MatBottomSheet, MatDialog } from '@angular/material';
 import { TestimonialsNewComponent } from './testimonials-new.component';
 import { Testimonial } from './testimonial';
 import { TestimonialService } from './testimonial.service';
@@ -8,6 +8,7 @@ import { TestimonialResponse } from './testimonial-response';
 import { HttpErrorResponse } from '@angular/common/http';
 import { catchError } from 'rxjs/operators';
 import { GlobalErrorHandlerService } from '../error/global-error-handler.service';
+import { DeleteConfirmationDialogComponent } from '../shared/delete-confirmation-dialog/delete-confirmation-dialog.component';
 
 @Component({
   selector: 'app-testimonials',
@@ -20,6 +21,7 @@ export class TestimonialsComponent implements OnInit {
   errorMessage: string = null;
 
   constructor(private testimonialService: TestimonialService,
+              public dialog: MatDialog,
               private changeDetector: ChangeDetectorRef,
               private globalErrorHandlerService: GlobalErrorHandlerService,
               private bottomSheet: MatBottomSheet) {
@@ -43,12 +45,20 @@ export class TestimonialsComponent implements OnInit {
   }
 
   delete(testimonial: TestimonialResponse) {
-    this.testimonialService.deleteTestimonial(testimonial.id)
-      .subscribe(() => {
-        this.loadTestimonials();
-      }, (errorResponse: HttpErrorResponse) => {
-       this.handleErrorResponse(errorResponse);
-      });
+    const dialogRef = this.dialog.open(DeleteConfirmationDialogComponent, {
+      width: '350px',
+      data: {}
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.testimonialService.deleteTestimonial(testimonial.id)
+          .subscribe(() => {
+            this.loadTestimonials();
+          }, (errorResponse: HttpErrorResponse) => {
+            this.handleErrorResponse(errorResponse);
+          });
+      }
+    });
   }
 
   private loadTestimonials() {
